Set the creation of admin on startup Only if it doesn't exists yet Set by env var Rework of the int tests seeds was required as it was overriding those values Int test are now sharing the seed with development Int test for projects are now without side effect between endpoints
