+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "build": "node client/build/build.js",
     "start": "concurrently --kill-others --success first \"yarn server-start\" \"yarn client-start\"",
-    "start-dev": "concurrently --kill-others --success first \"yarn server-start-dev --seed --trace-sync-io\" \"yarn client-start\"",
+    "start-dev": "concurrently --kill-others --success first \"yarn server-start-dev --trace-sync-io\" \"yarn client-start\"",
     "start-with-mocks": "concurrently --kill-others --success first \"yarn mock-server\" \"yarn client-start\" ",
     "test": "yarn client-unit && yarn server-unit && yarn server-int",
     "lint": "eslint --ext .js,.vue .",
