--- server/database/seed.js
@@ -1,4 +1,0 @@
-const path = require('path');
-
-module.exports = bookshelf =>
-  bookshelf.knex.seed.run({ directory: path.join(__dirname, '/seeds') });
