+++ server/test/integration/routes/projects.spec.js
@@ -1,33 +1,8 @@
 const { setup, cleanup } = require('../../setup-db');
 const request = require('supertest');
-const moment = require('moment');
 
 describe('integration: projects with open event by default', () => {
-  let token;
   let eventId;
-  let projectId;
-
-  before(async () => {
-    await setup();
-    return Promise.all([
-      getToken('owner@example.com')
-        .then((_token) => {
-          token = _token;
-          return Promise.resolve();
-        }),
-      getDefaultEvent(),
-    ]);
-  });
-
-  async function getToken(email) {
-    return request(app)
-      .post('/api/v1/users')
-      .set('Accept', 'application/json')
-      .send({ email, eventSlug: 'cp-2018' })
-      .then((res) => {
-        return Promise.resolve(res.body.auth.token);
-      });
-  }
 
   async function getDefaultEvent() {
     await request(app)
@@ -37,371 +12,21 @@ describe('integration: projects with open event by default', () => {
       });
   }
 
-  describe('/ post', () => {
-    it('should create a project', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        category: 'HTML',
-        users: [
-          {
-            firstName: 'kid1',
-            lastName: 'kido',
-            dob: '2008-12-01T20:00.000Z',
-            gender: 'male',
-            type: 'member',
-          },
-          {
-            firstName: 'orga',
-            lastName: 'le',
-            dob: '1991-12-01T20:00.000Z',
-            gender: 'male',
-            email: 'owner@example.com',
-            phone: '3538123123123',
-            country: 'IE',
-            type: 'supervisor',
-          },
-        ],
-      };
-      return request(app)
-        .post(`/api/v1/events/${eventId}/projects?token=${token}`)
-        .send(payload)
-        .expect('Content-Type', /json/)
-        .expect(200)
-        .then((res) => {
-          projectId = res.body.id;
-          expect(res.body).to.have.all.keys(['name', 'category', 'users', 'id', 'created_at', 'updated_at', 'eventId']);
-          expect(res.body.users.length).to.eql(2);
-          expect(res.body.users[0]).to.have.all.keys(['created_at', 'updated_at', 'id', 'firstName', 'lastName', 'type', 'dob', 'gender']);
-        });
-    });
-
-    it('should return an error when the eventId is invalid', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        category: 'HTML',
-        users: [
-          {
-            firstName: 'kid1',
-            lastName: 'kido',
-            dob: '2008-12-01T20:00.000Z',
-            gender: 'male',
-            type: 'member',
-          },
-          {
-            firstName: 'orga',
-            lastName: 'le',
-            dob: '1991-12-01T20:00.000Z',
-            gender: 'male',
-            email: 'owner@example.com',
-            phone: '3538123123123',
-            country: 'IE',
-            type: 'supervisor',
-          },
-        ],
-      };
-      return request(app)
-        .post(`/api/v1/events/861599a4-ca93-4a87-973f-276766f2a25b/projects?token=${token}`)
-        .send(payload)
-        .expect('Content-Type', /json/)
-        .expect(404)
-        .then((res) => {
-          expect(res.body.msg).to.equal('Event not found');
-          expect(res.body.status).to.equal(404);
-        });
-    });
-
-    it('should be requiering a valid auth', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        category: 'HTML',
-        users: [
-          {
-            firstName: 'kid1',
-            lastName: 'kido',
-            dob: '2008-12-01T20:00.000Z',
-            gender: 'male',
-            type: 'member',
-          },
-          {
-            firstName: 'orga',
-            lastName: 'le',
-            dob: '1991-12-01T20:00.000Z',
-            gender: 'male',
-            email: 'owner@example.com',
-            phone: '3538123123123',
-            country: 'IE',
-            type: 'supervisor',
-          },
-        ],
-      };
-      return request(app)
-        .post(`/api/v1/events/${eventId}/projects?token=aaa`)
-        .send(payload)
-        .expect(401);
-    });
-
-    it('should be allowing an email different for creator/supervisor', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        category: 'HTML',
-        users: [
-          {
-            firstName: 'kid1',
-            lastName: 'kido',
-            dob: '2008-12-01T20:00.000Z',
-            gender: 'male',
-            type: 'member',
-          },
-          {
-            firstName: 'orga',
-            lastName: 'le',
-            dob: '1991-12-01T20:00.000Z',
-            gender: 'male',
-            email: 'another@example.com',
-            phone: '3538123123123',
-            country: 'IE',
-            type: 'supervisor',
-          },
-        ],
-      };
-      return getToken('meme@example.com')
-        .then((_token) => {
-          return request(app)
-            .post(`/api/v1/events/${eventId}/projects?token=${_token}`)
-            .send(payload)
-            .expect(200);
-        });
-    });
-  });
-
-  describe('/:id PUT', () => {
-    const project = {
-      name: 'Self-overwriting-project',
-      category: 'Hardware',
-      users: [
-        {
-          firstName: 'replaceable',
-          lastName: 'Kid',
-          dob: '2008-12-01T20:00.000Z',
-          gender: 'male',
-          type: 'member',
-        },
-        {
-          firstName: 'exchangeable',
-          lastName: 'orga',
-          dob: '1991-12-01T20:00.000Z',
-          gender: 'male',
-          email: 'updatable@example.com',
-          phone: '3538123123123',
-          country: 'IE',
-          type: 'supervisor',
-        },
-      ],
-    };
-    let refProject;
-    let _token;
-
-    before(() => {
-      return util.user.create('updater@example.com')
-        .then((ntoken) => {
-          _token = ntoken;
-          return util.project.create(_token, eventId, project);
-        })
-        .then((res) => {
-          return util.project.get(_token, eventId, res.body.id);
-        })
-        .then((res) => { refProject = res.body; });
-    });
-
-    it('should update a project (same struct)', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.name = 'Self-updated';
-      payload.users[0].id = refProject.members[0].id;
-      payload.users[0].firstName = 'replaced';
-      payload.users[0].gender = 'female';
-      payload.users[1].id = refProject.supervisor.id;
-      await util.project.update(_token, eventId, refProject.id, payload)
-        .expect('Content-Type', /json/)
-        .expect(200)
-        .then((res) => {
-          expect(res.body).to.have.all.keys(['name', 'category', 'org', 'orgRef', 'state', 'city', 'description', 'answers', 'id', 'createdAt', 'updatedAt', 'deletedAt', 'eventId', 'users', 'status']);
-          expect(res.body.name).to.be.equal('Self-updated');
-          expect(res.body.users.length).to.eql(3);
-          const user = res.body.users.find(u => u.id === payload.users[0].id);
-          expect(user).to.have.all.keys(['createdAt', 'updatedAt', 'deletedAt', 'id', 'firstName', 'lastName', 'email', 'phone', 'country', 'specialRequirements', 'dob', 'gender']);
-          expect(user.firstName).to.be.equal('replaced');
-          expect(user.gender).to.be.equal('female');
-          expect(user.id).to.be.equal(payload.users[0].id);
-          const index = res.body.users.findIndex(u => u.id === refProject.supervisor.id);
-          expect(index).to.be.within(0, 3);
-        });
-    });
-
-    it('should be requiring a valid auth (owner)', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.users[0].id = refProject.members[0].id;
-      payload.users[1].id = refProject.supervisor.id;
-      await util.project.update(token, eventId, refProject.id, payload)
-        .expect('Content-Type', /json/)
-        .expect(403);
-    });
-    it('should allow change of emails', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.users[0].id = refProject.members[0].id;
-      payload.users[1].id = refProject.supervisor.id;
-      payload.users[1].email = 'updated@example.com';
-      await util.project.update(_token, eventId, refProject.id, payload)
-        .expect('Content-Type', /json/)
-        .expect(200)
-        .then((res) => {
-          expect(res.body.users.length).to.eql(3);
-          const user = res.body.users.find(u => u.id === payload.users[1].id);
-          expect(user.email).to.be.equal('updated@example.com');
-        });
-    });
-    it('should allow adding of users', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.users[0].id = refProject.members[0].id;
-      payload.users[1].id = refProject.supervisor.id;
-      payload.users[2] = {
-        firstName: 'freshAsNew',
-        lastName: 'Kid',
-        dob: '2008-12-01T20:00.000Z',
-        gender: 'other',
-        type: 'member',
-      };
-      await util.project.update(_token, eventId, refProject.id, payload)
-        .expect('Content-Type', /json/)
-        .expect(200)
-        .then((res) => {
-          expect(res.body.users.length).to.eql(4);
-          const newKid = res.body.users.find(u => u.firstName === payload.users[2].firstName);
-          expect([payload.users[0].id, payload.users[1].id]).to.not.include(newKid.id);
-        });
-    });
-    it('should allow removal of users', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.users.shift();
-      await util.project.update(_token, eventId, refProject.id, payload)
-        .expect('Content-Type', /json/)
-        .expect(200)
-        .then((res) => {
-          // original kid are deleted,
-          // only the owner and the supervisor are left
-          expect(res.body.users.length).to.equal(3);
-        });
-    });
-    it('should be disallowed any change if the event is frozen', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      payload.users[0].id = refProject.members[0].id;
-      payload.users[1].id = refProject.supervisor.id;
-      payload.users[1].email = 'updated@example.com';
-      const clock = sinon.useFakeTimers({
-        now: moment.utc().add(2, 'day').toDate(),
-      });
-
-      return util.user.create('updater@example.com')
-        .catch(() => {
-          return util.auth.get('updater@example.com')
-            .then((res) => {
-              return util.project.update(res.rows[0].token, eventId, refProject.id, payload)
-                .expect(403)
-                .then(() => {
-                  clock.restore();
-                });
-            });
-        });
-    });
-    it('should be allow any change if the event is open', async () => {
-      const payload = project;
-      project.id = refProject.id;
-      const clock = sinon.useFakeTimers({
-        now: moment.utc().add(1, 'day').set('hours', 0).toDate(),
-      });
-      return util.user.create('updater@example.com')
-        .catch(() => {
-          return util.auth.get('updater@example.com')
-            .then((res) => {
-              return util.project.update(res.rows[0].token, eventId, refProject.id, payload)
-                .expect(200)
-                .then(() => {
-                  clock.restore();
-                });
-            });
-        });
-    });
-  });
-
-  describe('/:id patch', () => {
-    it('should allow update of a project', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        answers: { social_project: true },
-      };
-      return request(app)
-        .patch(`/api/v1/events/${eventId}/projects/${projectId}?token=${token}`)
-        .send(payload)
-        .expect(200);
-    });
-    it('should fail if user is not the owner', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        answers: { social_project: true },
-      };
-      return getToken('anotherone@example.com')
-        .then((_token) => {
-          return request(app)
-            .patch(`/api/v1/events/${eventId}/projects/${projectId}?token=${_token}`)
-            .send(payload)
-            .expect(403)
-            .then((res) => {
-              expect(res.body.msg).to.equal('Forbidden');
-            });
-        });
-    });
-    it('should fail if event is frozen', async () => {
-      const payload = {
-        name: 'MyPoneyProject',
-        answers: { social_project: true },
-      };
-      const clock = sinon.useFakeTimers({
-        now: moment.utc().add(1, 'day').toDate(),
-      });
-      return util.user.create('updater@example.com')
-        .catch(() => {
-          return util.auth.get('updater@example.com')
-            .then((res) => {
-              return request(app)
-                .patch(`/api/v1/events/${eventId}/projects/${projectId}?token=${res.rows[0].token}`)
-                .send(payload)
-                .expect(403)
-                .then(() => {
-                  clock.restore();
-                });
-            });
-        });
-    });
-  });
-
   describe('/users/:uId/projects GET', () => {
     let userId;
-    let refProject;
     let refToken;
-    before(() => {
-      return util.user.create('listuserprojects@example.com', true)
+    let refProject;
+    before(async () => {
+      await setup();
+      await getDefaultEvent();
+      await util.user.create('listuserprojects@example.com', true)
         .then((res) => {
           userId = res.user.id;
           refToken = res.auth.token;
-        })
-        .then(() => util.project.create(refToken, eventId))
-        .then((res) => { refProject = res.body; });
+        });
+      refProject = (await util.project.create(refToken, eventId)).body;
     });
+    after(cleanup);
     it('should return the user projects', async () => {
       await request(app)
         .get(`/api/v1/events/${eventId}/users/${userId}/projects?token=${refToken}`)
@@ -451,283 +76,4 @@ describe('integration: projects with open event by default', () => {
         });
     });
   });
-
-  describe('/ GET', () => {
-    let refAuth;
-    let firstTenProjects;
-    const count = 55;
-    before(() => {
-      return Promise.all([
-        util.auth.get('hello@coolestprojects.org')
-          .then((res) => {
-            refAuth = res.rows[0];
-          }),
-        (async () => {
-          const projects = new Array(50);
-          // eslint-disable-next-line no-restricted-syntax
-          for (const i of projects.keys()) {
-            // eslint-disable-next-line no-await-in-loop
-            await util.user.create(`test${i}@test.com`)
-              // eslint-disable-next-line no-loop-func
-              .then(_token => util.project.create(_token, eventId, {
-                name: `MyPoneyProject${i}`,
-                category: 'HTML',
-                users: [
-                  {
-                    firstName: 'kid1',
-                    lastName: 'kido',
-                    dob: '2008-12-01T20:00.000Z',
-                    gender: 'male',
-                    type: 'member',
-                  },
-                  {
-                    firstName: 'orga',
-                    lastName: 'le',
-                    dob: '1991-12-01T20:00.000Z',
-                    gender: 'male',
-                    email: `test${i}@example.com`,
-                    phone: '3538123123123',
-                    country: 'IE',
-                    type: 'supervisor',
-                  },
-                ],
-              }));
-          }
-        })(),
-      ]);
-    });
-
-    it('should return by default a list of 25 projects ordered by creation date desc', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          // Content
-          expect(res.body.count).to.equal(count);
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.data[0]).to.have.all.keys(['id', 'name', 'category', 'createdAt', 'updatedAt', 'deletedAt', 'eventId', 'description', 'answers', 'org', 'orgRef', 'state', 'city', 'owner', 'supervisor', 'members', 'status']);
-          expect(res.body.data[0].owner).to.have.all.keys(['id', 'firstName', 'lastName', 'dob', 'gender', 'specialRequirements', 'email', 'phone', 'country', 'createdAt', 'updatedAt', 'deletedAt']);
-          expect(res.body.data[0].supervisor).to.have.all.keys(['id', 'firstName', 'lastName', 'dob', 'gender', 'specialRequirements', 'email', 'phone', 'country', 'createdAt', 'updatedAt', 'deletedAt']);
-          expect(res.body.data[0].members[0]).to.have.all.keys(['id', 'firstName', 'lastName', 'dob', 'gender', 'specialRequirements', 'email', 'phone', 'country', 'createdAt', 'updatedAt', 'deletedAt']);
-          // Ordering
-          expect(res.body.data.length).to.equal(25);
-          expect(res.body.data[0].name).to.eql('MyPoneyProject49');
-        });
-    });
-
-    it('should return a list of 10 projects when provided a limit', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=10&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.count).to.equal(count);
-          expect(res.body.data.length).to.equal(10);
-          firstTenProjects = res.body.data;
-        });
-    });
-
-    it('should return a list of 10 projects different on page 2 when provided a limit and a page', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=10&page=2&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.count).to.equal(count);
-          expect(res.body.data.length).to.equal(10);
-          expect(res.body.data).not.to.be.eql(firstTenProjects);
-          expect(res.body.data.map(project => project.name))
-            .not.to.be.eql(firstTenProjects.map(project => project.name));
-        });
-    });
-    it('should order the list depending on query', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=20&orderBy=name&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.count).to.equal(count);
-          expect(res.body.data.length).to.equal(20);
-          expect(res.body.data[0].name).to.be.eql('Self-updated');
-        });
-    });
-    it('should allow ordering by subqueries values (owner)', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=10&orderBy=owner.email&ascending=1&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.count).to.equal(count);
-          // TODO : better define a project just for this one
-          // which starts with '0aaaa' to ensure the test always passes
-          expect(res.body.data[0].owner.email).to.be.eql('listuserprojects@example.com');
-          expect(res.body.data[0].name).to.be.eql('MyPoneyProject');
-        });
-    });
-    it('should allow ordering by subqueries values (supervisor)', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          expect(res.body.count).to.equal(count);
-          expect(res.body.data.length).to.equal(50);
-          expect(res.body.data[0].supervisor.email).to.be.eql('updated@example.com');
-          expect(res.body.data[0].name).to.be.eql('Self-updated');
-        });
-    });
-    it('should allow filtering', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&query[name]=Project1&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          // Project from 1 to 10-19
-          expect(res.body.count).to.equal(11);
-          expect(res.body.data.length).to.equal(11);
-        });
-    });
-    it('should allow filtering from sub-orgs', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&query[supervisor.email]=test1&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          // Emails from 1 to 10-19
-          expect(res.body.count).to.equal(11);
-          expect(res.body.data.length).to.equal(11);
-        });
-    });
-    it('should allow multiple filtering', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&query[name]=MyPoneyProject&query[supervisor.email]=another&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          // Project with supervisor from prev test
-          expect(res.body.count).to.equal(1);
-          expect(res.body.data.length).to.equal(1);
-        });
-    });
-    it('should ignore empty filters', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&query[category]=&query[name]=MyPoneyProject&query[supervisor.email]=another&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          // Project with supervisor from prev test
-          expect(res.body.count).to.equal(1);
-          expect(res.body.data.length).to.equal(1);
-        });
-    });
-    it('should support camelCase names', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=10&orderBy=supervisor.createdAt&query[orgRef]=choubidou&ascending=0&token=${refAuth.token}`)
-        .expect(200)
-        .then((res) => {
-          expect(Object.keys(res.body)).to.eql(['data', 'count']);
-          // Project with supervisor from prev test
-          expect(res.body.count).to.equal(0);
-          expect(res.body.data.length).to.equal(0);
-        });
-    });
-    it('should be covered by admin token', async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?token=${token}`)
-        .expect(403);
-    });
-
-    it("shouldn't include deleted projects", async () => {
-      await request(app)
-        .get(`/api/v1/events/${eventId}/projects?limit=1&token=${refAuth.token}`)
-        .expect(200)
-        .then(async (res) => {
-          const lastProject = res.body.data[0];
-          await util.project.delete(refAuth.token, eventId, lastProject);
-          await request(app)
-            .get(`/api/v1/events/${eventId}/projects?limit=1&token=${refAuth.token}`)
-            .expect(200)
-            .then((r) => {
-              expect(r.body.data[0].id).to.not.equal(lastProject.id);
-            });
-        });
-    });
-    describe('with csv format', () => {
-      const projectCSVColumns = [
-        '"Project name"',
-        '"Description"',
-        '"Category"',
-        '"City"',
-        '"State"',
-        '"Owner Email"',
-        '"Seat"',
-        '"Status"',
-        '"Created At"',
-        '"Updated At"',
-        '"Supervisor First Name"',
-        '"Supervisor Last Name"',
-        '"Supervisor Email"',
-        '"Supervisor Phone"',
-        '"Social project"',
-        '"Educational project"',
-        '"Innovator stage"',
-      ];
-
-      const csvColumnsWithParticipant = projectCSVColumns.concat([
-        '"Participant 1 First Name"',
-        '"Participant 1 Last Name"',
-        '"Participant 1 Dob"',
-        '"Participant 1 Gender"',
-        '"Participant 1 Special requirements"',
-      ]);
-      const userCSVColumns = ['"First name"', '"Last name"', '"Dob"', '"Gender"', '"Special requirements"'].concat(projectCSVColumns);
-      it('should return a csv', async () => {
-        await request(app)
-          .get(`/api/v1/events/${eventId}/projects?limit=50&orderBy=supervisor.email&query[supervisor.email]=test1&format=csv&ascending=false&token=${
-            refAuth.token
-          }`)
-          .expect(200)
-          .then((res) => {
-            expect(res.text).not.to.be.empty;
-            const lines = res.text.split('\n');
-            // 10-19 + 1 + headers due to supervisor email filtering
-            expect(lines.length).to.equal(12);
-            const columns = lines[0].split(',');
-            const row = lines[1].split(',');
-            expect(columns).to.eql(csvColumnsWithParticipant);
-            expect(row[row.length - 13]).to.eql(`"${new Date(Date.now()).toLocaleDateString()}"`);
-          });
-      });
-      it('should return an empty csv with headers', async () => {
-        await request(app)
-          .get(`/api/v1/events/${eventId}/projects?limit=50&query[supervisor.email]=doubidou&format=csv&ascending=false&token=${
-            refAuth.token
-          }`)
-          .expect(200)
-          .then((res) => {
-            expect(res.text).not.to.be.empty;
-            const lines = res.text.split('\n');
-            expect(lines.length).to.equal(1); // headers
-            const columns = lines[0].split(',');
-            expect(columns).to.eql(projectCSVColumns);
-          });
-      });
-      describe('with view === user', () => {
-        it('should return the projects with the a user per row', async () => {
-          await request(app)
-            .get(`/api/v1/events/${eventId}/projects?view=user&format=csv&limit=50&token=${refAuth.token}`)
-            .expect(200)
-            .then((res) => {
-              expect(res.text).not.to.be.empty;
-              const lines = res.text.split('\n');
-              // Limit is "wrong" because it's only applied to the number of project
-              expect(lines.length).to.equal(54); // 52 + header
-              const columns = lines[0].split(',');
-              expect(columns).to.eql(userCSVColumns);
-            });
-        });
-      });
-    });
-  });
-  after(cleanup);
 });
