+++ server/test/integration/routes/users.spec.js
@@ -1,4 +1,3 @@
-// TODO fix parallel tests
 const { setup, cleanup } = require('../../setup-db');
 const request = require('supertest');
 
@@ -72,8 +71,8 @@ describe('integration: users', () => {
       await util.user.get(adminToken)
         .expect(200)
         .then((res) => {
-          expect(res.body.count).to.equal(4);
-          expect(res.body.data.length).to.equal(4);
+          expect(res.body.count).to.equal(333);
+          expect(res.body.data.length).to.equal(25);
           expect(res.body.data[0]).to.have.all.keys(['id', 'firstName', 'lastName', 'dob', 'gender', 'specialRequirements', 'email', 'phone', 'country', 'createdAt', 'updatedAt', 'deletedAt', 'membership']);
         });
     });
@@ -88,7 +87,7 @@ describe('integration: users', () => {
       await util.user.get(adminToken, 'query[gender]=female')
         .expect(200)
         .then((res) => {
-          expect(res.body.count).to.equal(0);
+          expect(res.body.count).to.be.below(110 * 3);
         });
     });
   });
